Reorganize Features tab components into separate files
