+++ static/js/components/UploadFeaturesForm.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+
+const UploadFeaturesForm = props => {
+  return (
+    <div>
+      Upload Features Form content
+    </div>
+  );
+};
+
+
+export default UploadFeaturesForm;
